requirements.txt
aiosqlite==0.16.0
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
databases==0.4.1
dnspython==2.0.0
email-validator==1.1.2
fastapi==0.62.0
fastapi-login==1.5.1
fastapi-users==3.1.1
h11==0.11.0
idna==2.10
makefun==1.9.5
passlib==1.7.4
pycparser==2.20
pydantic==1.7.3
PyJWT==1.7.1
python-dotenv==0.15.0
python-multipart==0.0.5
six==1.15.0
SQLAlchemy==1.3.20
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.0
